Migration assistance (#3445)
* Log errors when the PostgreSQL data directory is wrong

The postgres-startup container now reports when it finds the installed
PostgreSQL binaries do not match the specified PostgreSQL version.

Some storage providers do not mount the PostgreSQL data volume with
correct ownership or permissions. The postgres-startup container now
prints those attributes of parent directories when it cannot create or
modify a needed file or directory.

Issue: [sc-11804]
Issue: CrunchyData/postgres-operator#2870


* Change owner of the PostgreSQL directory at startup
PostgreSQL won't to start unless it owns the data directory. Kubernetes
sets the group according to fsGroup but not the owner.

The postgres-startup container now recreates the data directory to give
it a new owner when permissions are sufficient to do so. It now raises
an error when the owner is incorrect and cannot be changed.

Issue: [sc-15909]
See: https://docs.k8s.io/tasks/configure-pod-container/security-context/


* Add KUTTL test for migration from third-party PGSQL

Issue: [sc-15909]

+++ internal/postgres/reconcile_test.go
@@ -202,7 +202,13 @@ initContainers:
   - --
   - |-
     declare -r expected_major_version="$1" pgwal_directory="$2" pgbrLog_directory="$3"
+    permissions() { while [[ -n "$1" ]]; do set "${1%/*}" "$@"; done; shift; stat -Lc '%A %4u %4g %n' "$@"; }
+    halt() { local rc=$?; >&2 echo "$@"; exit "${rc/#0/1}"; }
     results() { printf '::postgres-operator: %s::%s\n' "$@"; }
+    recreate() (
+      local tmp; tmp=$(mktemp -d -p "${1%/*}"); GLOBIGNORE='.:..'; set -x
+      chmod "$2" "${tmp}"; mv "$1"/* "${tmp}"; rmdir "$1"; mv "${tmp}" "$1"
+    )
     safelink() (
       local desired="$1" name="$2" current
       current=$(realpath "${name}")
@@ -214,21 +220,32 @@ initContainers:
     results 'uid' "$(id -u)" 'gid' "$(id -G)"
     results 'postgres path' "$(command -v postgres)"
     results 'postgres version' "${postgres_version:=$(postgres --version)}"
-    [[ "${postgres_version}" == *") ${expected_major_version}."* ]]
+    [[ "${postgres_version}" =~ ") ${expected_major_version}"($|[^0-9]) ]] ||
+    halt Expected PostgreSQL version "${expected_major_version}"
     results 'config directory' "${PGDATA:?}"
     postgres_data_directory=$([ -d "${PGDATA}" ] && postgres -C data_directory || echo "${PGDATA}")
     results 'data directory' "${postgres_data_directory}"
-    [ "${postgres_data_directory}" = "${PGDATA}" ]
+    [[ "${postgres_data_directory}" == "${PGDATA}" ]] ||
+    halt Expected matching config and data directories
     bootstrap_dir="${postgres_data_directory}_bootstrap"
     [ -d "${bootstrap_dir}" ] && results 'bootstrap directory' "${bootstrap_dir}"
     [ -d "${bootstrap_dir}" ] && postgres_data_directory="${bootstrap_dir}"
+    if [[ ! -e "${postgres_data_directory}" || -O "${postgres_data_directory}" ]]; then
     install --directory --mode=0700 "${postgres_data_directory}"
+    elif [[ -w "${postgres_data_directory}" && -g "${postgres_data_directory}" ]]; then
+    recreate "${postgres_data_directory}" '0700'
+    else (halt Permissions!); fi ||
+    halt "$(permissions "${postgres_data_directory}" ||:)"
     results 'pgBackRest log directory' "${pgbrLog_directory}"
-    install --directory --mode=0775 "${pgbrLog_directory}"
+    install --directory --mode=0775 "${pgbrLog_directory}" ||
+    halt "$(permissions "${pgbrLog_directory}" ||:)"
     install -D --mode=0600 -t "/tmp/replication" "/pgconf/tls/replication"/{tls.crt,tls.key,ca.crt}
     [ -f "${postgres_data_directory}/PG_VERSION" ] || exit 0
     results 'data version' "${postgres_data_version:=$(< "${postgres_data_directory}/PG_VERSION")}"
-    [ "${postgres_data_version}" = "${expected_major_version}" ]
+    [[ "${postgres_data_version}" == "${expected_major_version}" ]] ||
+    halt Expected PostgreSQL data version "${expected_major_version}"
+    [[ ! -f "${postgres_data_directory}/postgresql.conf" ]] &&
+    touch "${postgres_data_directory}/postgresql.conf"
     safelink "${pgwal_directory}" "${postgres_data_directory}/pg_wal"
     results 'wal directory' "$(realpath "${postgres_data_directory}/pg_wal")"
     rm -f "${postgres_data_directory}/recovery.signal"
